requirements.txt
fastapi
uvicorn[standard] # For running the FastAPI app
requests
pandas
scikit-learn
python-dotenv # For loading environment variables
psycopg2-binary # If using PostgreSQL